refactor(todos): streamline todo input handling and enhance context integration
