+++ src/layouts/widgets/todos/expandable-todo-input.tsx
@@ -7,6 +7,7 @@ import {
 	getTextColor,
 	useTheme,
 } from '@/context/theme.context'
+import { type AddTodoInput, TodoPriority } from '@/context/todo.context'
 import { AnimatePresence, motion } from 'framer-motion'
 import { useEffect, useRef, useState } from 'react'
 import { FiFlag, FiMessageSquare, FiPlus, FiTag, FiX } from 'react-icons/fi'
@@ -14,12 +15,7 @@ import { FiFlag, FiMessageSquare, FiPlus, FiTag, FiX } from 'react-icons/fi'
 interface ExpandableTodoInputProps {
 	todoText: string
 	onChangeTodoText: (value: string) => void
-	onAddTodo: (
-		text: string,
-		priority: 'low' | 'medium' | 'high',
-		category?: string,
-		notes?: string,
-	) => void
+	onAddTodo: (input: Omit<AddTodoInput, 'date'>) => void
 }
 
 const PrIORITY_OPTIONS = [
@@ -50,7 +46,7 @@ export function ExpandableTodoInput({
 }: ExpandableTodoInputProps) {
 	const { theme } = useTheme()
 	const [isExpanded, setIsExpanded] = useState(false)
-	const [priority, setPriority] = useState<'low' | 'medium' | 'high'>('medium')
+	const [priority, setPriority] = useState<TodoPriority>(TodoPriority.Medium)
 	const [category, setCategory] = useState('')
 	const [notes, setNotes] = useState('')
 	const inputRef = useRef<HTMLInputElement | null>(null)
@@ -81,7 +77,12 @@ export function ExpandableTodoInput({
 
 	const handleAddTodo = () => {
 		if (todoText.trim()) {
-			onAddTodo(todoText.trim(), priority, category, notes)
+			onAddTodo({
+				text: todoText.trim(),
+				category: category.trim() || undefined,
+				notes: notes.trim() || undefined,
+				priority: priority,
+			})
 			Analytics.featureUsed('todo_added')
 			resetForm()
 		}
@@ -97,7 +98,7 @@ export function ExpandableTodoInput({
 		onChangeTodoText('')
 		setCategory('')
 		setNotes('')
-		setPriority('medium')
+		setPriority(TodoPriority.Medium)
 		setIsExpanded(false)
 	}
 
@@ -165,9 +166,7 @@ export function ExpandableTodoInput({
 													<button
 														key={value}
 														type="button"
-														onClick={() =>
-															setPriority(value as 'low' | 'medium' | 'high')
-														}
+														onClick={() => setPriority(value as TodoPriority)}
 														className={`
 													flex items-center justify-center w-4 h-4 rounded-full
 													transition-all duration-150 cursor-pointer
